Correct bug in bitmap reader Switch to 64-bit image

+++ HyperCardCommon/HyperCardCommon/Image.swift
@@ -14,7 +14,7 @@ import AppKit
 /// We don't use a Cocoa image because there are a lot of specific processes in 1-bit images.
 public struct Image {
     
-    public typealias Integer = UInt32
+    public typealias Integer = UInt64
     
     /// The bits, stored in 32-bit integers. A row always starts at the beginning of an integer,
     /// so there may be unused bits at the end of every row for 32-bit alignment.
